Made Model Building tests folder and added to tests __init__
