+++ DRP/tests/__init__.py
@@ -5,20 +5,20 @@ import Email
 import forms
 import HttpTests
 import decorators
-import ModelContainer
+import ModelBuildingTests
 import CompoundFromCsv
 import CompoundDescriptor
 import CompoundToCsv
 import CompoundToArff
 
 suite = unittest.TestSuite([
-  Email.suite,
-  forms.suite,
-  HttpTests.suite,
-  ModelContainer.suite,
-  CompoundFromCsv.suite,
-  CompoundDescriptor.suite,
-  CompoundToCsv.suite,
-  CompoundToArff.suite,
-  fileTests.suite,
+    Email.suite,
+    forms.suite,
+    HttpTests.suite,
+    ModelBuildingTests.suite,
+    CompoundFromCsv.suite,
+    CompoundDescriptor.suite,
+    CompoundToCsv.suite,
+    CompoundToArff.suite,
+    fileTests.suite,
 ])
